Bump PLATFORM_VERSION to 3.4.0
To allow a system upgrade, the installer checks and compares the values
of PLATFORM_VERSION set by the branding package.

In order to allow an upgrade from XenServer 8.4 (PLATFORM_VERSION=3.4.0)
to XCP-ng 8.3, this change sets PLATFORM_VERSION to 3.4.0. That way
XCP-ng can stick to PRODUCT_VERSION=8.3 and allow such upgrades without
modifying the installer.

Signed-off-by: Thierry Escande <thierry.escande@vates.tech>