Switch product version to 8.99.0, platform XCPNG-9.0.0
The platform name itself swtches from "XCP" to "XCPNG" since "XCP" is
the name of the platform used by XenServer, and we're in the process
of using a different OS base, with potentially different choices.

At the same time reorder minor/micro variables in a sane order.

Continuing using this package as is really a temporary measure to
allow building in koji while we get the bitbake-based system into
shape.  Its contents will be dispatched in more logical places soon.

Signed-off-by: Yann Dirson <yann.dirson@vates.tech>